requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==1.0.0
Django==4.0.6
django-extensions==3.2.0
django-on-heroku==1.1.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-decouple==3.6
python-dotenv==0.20.0
sqlparse==0.4.2
whitenoise==6.2.0
